servicekit/requirements.txt
# appdirs==1.4.0

backports.shutil-get-terminal-size==1.0.0
cloudconvert==1.0.0

Django==1.9
django-easy-select2==1.3.3
django-filebrowser-no-grappelli==3.6.2
django-formtools==1.0
# django-grappelli==2.8.1
django-model-utils==2.6.1
django-sortedm2m==1.3.2
# django-taggit==0.20.2
# docutils==0.12
# docx-mailmerge==0.3.0
# docxtpl==0.2.2
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
# Jinja2==2.8
jsonfield==1.0.3
lxml==3.6.1
# MarkupSafe==0.23
# packaging==16.8
pathlib2==2.1.0
enum34==1.1.6


Pillow==3.3.1
# pockets==0.3


# Pygments==2.1.3
# pyparsing==2.1.10
PyPDF2==1.26.0
# python-docx==0.8.6
reportlab==3.4.0
requests==2.13.0


# Sphinx==1.2.3
# sphinxcontrib-napoleon==0.5.2


# xlrd==1.0.0
# xlwt==1.1.2


